+++ backend/src/Controller/User/CharacterController.php
@@ -102,7 +102,7 @@ class CharacterController extends BaseController
      *     path="/user/character/find-by/{name}",
      *     operationId="findBy",
      *     summary="Return a list of characters that matches the name (partial matching).",
-     *     description="Needs role: user-admin or group-manager",
+     *     description="Needs role: user-admin, user-manager, group-manager",
      *     tags={"Character"},
      *     security={{"Session"={}}},
      *     @OA\Parameter(