+++ frontend/src/pages/PlayerGroupManagement.vue
@@ -13,14 +13,28 @@
                     <a :href="httpBaseUrl + '/login-managed-alt'">{{ httpBaseUrl }}/login-managed-alt</a>
                 </p>
                 <p class="small text-muted">
-                    This login URL does not require ESI scopes. This disables groups for the player 
-                    account (if this feature is enabled) unless the status is "managed".
+                    These login URLs do not require ESI scopes. They disable groups for the player account,
+                    if the "Deactivate Groups" feature is enabled, unless their account status is "managed".
                 </p>
             </div>
         </div>
 
         <div class="row">
             <div class="col-lg-4">
+                <div class="card border-secondary mb-3" >
+                    <h3 class="card-header">Characters</h3>
+                    <div class="card-body">
+                        <character-search v-on:result="onSearchResult($event)"></character-search>
+                    </div>
+                    <div class="list-group">
+                        <a v-for="char in searchResult"
+                           class="list-group-item list-group-item-action"
+                           :class="{ active: playerId === char.player_id }"
+                           :href="'#PlayerGroupManagement/' + char.player_id">
+                            {{ char.character_name }}
+                        </a>
+                    </div>
+                </div>
                 <div class="card border-secondary mb-3" >
                     <h3 class="card-header">
                         Players
@@ -39,25 +53,42 @@
             </div>
             <div class="col-lg-8">
                 <div class="card border-secondary" >
-                    <h3 class="card-header">
-                        Groups
-                    </h3>
+                    <h3 class="card-header">Groups</h3>
                     <div v-if="playerData" class="card-body">
-                        <span class="text-muted">
+                        <h4>
                             <span v-if="hasRole('user-admin')">
-                                <a :href="'#UserAdmin/' + playerData.id">{{ playerData.name }} #{{ playerData.id }}</a>,
+                                <a :href="'#UserAdmin/' + playerData.id">{{ playerData.name }} #{{ playerData.id }}</a>
                             </span>
-                            <span v-if="! hasRole('user-admin')">{{ playerData.name }} #{{ playerData.id }},</span>
-                            status: {{ playerData.status }}
-                        </span>
-                        <a class="badge badge-info ml-1" href="" v-on:click.prevent="showCharacters(playerData.id)">
-                            Show characters
-                        </a>
-                        <span v-if="playerData.status === 'standard'" class="text-warning">
-                            <br>
-                            The status of this player is not "managed", manual changes can
+                            <span v-if="! hasRole('user-admin')">{{ playerData.name }} #{{ playerData.id }}</span>
+                        </h4>
+                        <p>
+                            Status: {{ playerData.status }}
+                            <a class="badge badge-info ml-1" href="" v-on:click.prevent="showCharacters(playerData.id)">
+                                Show characters
+                            </a>
+                        </p>
+                        <p v-if="playerData.status === 'standard'" class="text-warning">
+                            The status of this player is not "managed", manual changes will
                             be overwritten by the automatic group assignment.
-                        </span>
+                        </p>
+
+                        <hr>
+
+                        <h5>Account Status</h5>
+                        <p class="text-warning">
+                            All groups will be removed from the player account when the status is changed!
+                        </p>
+                        <div class="input-group mb-1">
+                            <div class="input-group-prepend">
+                                <label class="input-group-text" for="userAdminSetStatus">status</label>
+                            </div>
+                            <select class="custom-select" id="userAdminSetStatus"
+                                    v-model="playerData.status"
+                                    v-on:change="setAccountStatus()">
+                                <option value="standard">standard</option>
+                                <option value="managed">manually managed</option>
+                            </select>
+                        </div>
                     </div>
                 </div>
 
@@ -73,14 +104,16 @@
 </template>
 
 <script>
-import { PlayerApi } from 'neucore-js-client';
-import Admin      from '../components/EntityRelationEdit.vue';
-import Characters from '../components/Characters.vue';
+import { PlayerApi }   from 'neucore-js-client';
+import Admin           from '../components/EntityRelationEdit.vue';
+import Characters      from '../components/Characters.vue';
+import CharacterSearch from '../components/CharacterSearch.vue';
 
 export default {
     components: {
         Admin,
         Characters,
+        CharacterSearch,
     },
 
     props: {
@@ -95,6 +128,7 @@ export default {
             playerId: null, // current player
             playerData: null, // current player
             httpBaseUrl: null,
+            searchResult: [],
         }
     },
 
@@ -106,7 +140,7 @@ export default {
 
         // login URL for managed accounts
         let port = '';
-        if (location.port !== "" && location.port !== 80 && location.port !== 443) {
+        if (location.port !== "" && location.port + '' !== '80' && location.port + '' !== '443') {
             port = ':' + location.port;
         }
         this.httpBaseUrl = location.protocol + "//" + location.hostname + port
@@ -135,6 +169,26 @@ export default {
 
         showCharacters: function(memberId) {
             this.$refs.charactersModal.showCharacters(memberId);
+        },
+
+        onSearchResult: function(result) {
+            this.searchResult = result;
+        },
+
+        setAccountStatus: function() {
+            const vm = this;
+            const playerId = vm.playerId;
+            new PlayerApi().setStatus(playerId, vm.playerData.status, function(error) {
+                if (error) {
+                    return;
+                }
+                vm.getPLayers();
+                vm.$refs.admin.getSelectContent();
+                vm.$refs.admin.getTableContent();
+                if (playerId === vm.player.id) {
+                    vm.$root.$emit('playerChange');
+                }
+            });
         },
     },
 }