+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Change: Added role watchlist-manager with permissions to edit the configuration, 
   edit access is now restricted to watchlist-admin.
+- Change: Moved "set account status" function from the role user-admin to user-manager.
 - Some refactoring
 
 ## 1.9.2